refactor: use tsconfig aliases
