+++ services/api/bitget/data-transformer.ts
@@ -9,8 +9,8 @@
  */
 
 import { ExchangeType } from '@/types/api';
-import { OHLCData, Timeframe } from '../../../types/chart';
-import { OrderBookData } from '../../../types/market';
+import { OHLCData, Timeframe } from '@/types/chart';
+import { OrderBookData } from '@/types/market';
 import { logger } from '@/utils/common';
 import { IS_DEV } from '@/config/environment';
