+++ services/api/bitget/interfaces.ts
@@ -9,8 +9,8 @@
  */
 
 import { ProductType, BitgetCredentials } from '@/types/api';
-import { OHLCData } from '../../../types/chart';
-import { OrderBookData } from '../../../types/market';
+import { OHLCData } from '@/types/chart';
+import { OrderBookData } from '@/types/market';
 import { IRestApiClient, IWebSocketClient } from '../interfaces';
 
 /**
