+++ services/api/bitget/rest-client.ts
@@ -12,7 +12,7 @@
 
 import axios, { AxiosError } from 'axios';
 import { ExchangeType, ExchangeProductType } from '@/types/api';
-import { OHLCData, Timeframe, OrderBookData } from '../../../types/chart';
+import { OHLCData, Timeframe, OrderBookData } from '@/types/chart';
 import { IRestApiClient } from '../interfaces';
 import { logger } from '../../../utils/logger';
 import { safeExchangeType, safeProductType } from '../../../utils/exchangeTypeUtils';
