+++ services/api/chart-data-service.ts
@@ -11,7 +11,7 @@
 
 import { EventEmitter } from 'events';
 import { ExchangeType } from '@/types/api';
-import { OHLCData, OrderBookData, Timeframe } from '../../types/chart';
+import { OHLCData, OrderBookData, Timeframe } from '@/types/chart';
 import { IChartDataService } from './interfaces';
 import { dataSourceFactory } from './data-source-factory';
 import { logger } from '../../utils/logger';
