+++ services/data/order-book-service.ts
@@ -9,7 +9,7 @@
 import { EventEmitter } from 'events';
 import { BitgetApiClient } from '../api/bitget/client.new';
 import { IOrderBookService } from '../api/interfaces';
-import { OrderBookData } from '../../types/chart';
+import { OrderBookData } from '@/types/chart';
 import { ExchangeType } from '@/types/api';
 import { logger } from '../../utils/logger';
 import { normalizeSymbol } from '../../utils/formatters';
