+++ services/socket/interfaces.ts
@@ -7,8 +7,8 @@
  */
 
 import { Socket } from 'socket.io-client';
-import { OrderBookData } from '../../types/market';
-import { OHLCData, Timeframe } from '../../types/chart';
+import { OrderBookData } from '@/types/market';
+import { OHLCData, Timeframe } from '@/types/chart';
 import { ProductType } from '@/types/api';
 
 /**
