+++ src/mastra/tools/symbol-tools.ts
@@ -7,9 +7,9 @@
 
 import { createTool } from "@mastra/core/tools";
 import { z } from "zod";
-import { logger } from "../../../utils/common";
+import { logger } from "@/utils/common";
 import fetch from "node-fetch";
-import { ExchangeType } from "../../../types/api";
+import { ExchangeType } from "@/types/api";
 
 /**
  * チャートの銘柄を変更するツール
