+++ src/mastra/tools/timeframe-tools.ts
@@ -4,7 +4,7 @@
 
 import { createTool } from "@mastra/core/tools";
 import { z } from "zod";
-import { logger } from "../../../utils/common";
+import { logger } from "@/utils/common";
 import fetch from "node-fetch";
 
 /**
